+++ src/raw_bmp.rs
@@ -44,18 +44,32 @@ impl<'a> RawBmp<'a> {
 
         let color_type = try_const!(ColorType::from_header(&header));
 
-        // Believe what the bitmap tells us rather than multiplying width by
-        // height by bits-per-pixel, because the image data might be compressed.
-        let compressed_data_length = header.image_data_len as usize;
-
         if bytes.len() < header.image_data_start {
             return Err(ParseError::UnexpectedEndOfFile);
         }
         let (_, image_data) = bytes.split_at(header.image_data_start);
-        if image_data.len() < compressed_data_length {
+
+        let data_length = if let crate::header::CompressionMethod::Rgb = header.compression_method {
+            // `Header::image_data_len` may be zero or bogus when compression mode is RGB
+            // see `biSizeImage` on https://learn.microsoft.com/en-us/previous-versions/dd183376(v=vs.85)
+            // so we should calculate width x height instead.
+            let height = header.image_size.height as usize;
+
+            let Some(data_length) = header.bytes_per_row().checked_mul(height) else {
+                return Err(ParseError::UnexpectedEndOfFile);
+            };
+            data_length
+        } else {
+            // Believe what the bitmap tells us rather than multiplying width by
+            // height by bits-per-pixel, because the image data might be compressed.
+            header.image_data_len as usize
+        };
+
+        if image_data.len() < data_length {
             return Err(ParseError::UnexpectedEndOfFile);
         }
-        let (image_data, _) = image_data.split_at(compressed_data_length);
+
+        let (image_data, _) = image_data.split_at(data_length);
 
         Ok(Self {
             header,