+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - [#46](https://github.com/embedded-graphics/tinybmp/pull/46) `Bmp::from_slice` is now `const`, so BMPs can be put in `const`s and `static`s.
 
+- [#47](https://github.com/embedded-graphics/tinybmp/pull/47) Ignore compressed data length on RGB compression mehod
 
 ## [0.6.0] - 2024-06-11